requirements.txt
eth-ape
ape-etherscan
ape-alchemy
pycoingecko
altair